+++ src/detections/field_data_map.rs
@@ -133,7 +133,10 @@ fn load_yaml_files(dir_path: &Path) -> Result<Vec<Yaml>, String> {
             .flatten()
             .collect()),
         Err(e) => {
-            let msg = format!("Failed to open field mapping dir[{path}].");
+            let mut msg = format!("Failed to open field mapping dir[{path}]. ",);
+            if e.to_string().ends_with("123)") {
+                msg = format!("{msg}. You may not be able to load evtx files when there are spaces in the directory path. Please enclose the path with double quotes and remove any trailing slash at the end of the path.");
+            }
             AlertMessage::warn(&msg).ok();
             Err(e.to_string())
         }