+++ src/main.rs
@@ -881,7 +881,10 @@ impl App {
         }
         let entries = fs::read_dir(dirpath);
         if entries.is_err() {
-            let errmsg = format!("{}", entries.unwrap_err());
+            let mut errmsg = format!("{}", entries.unwrap_err());
+            if errmsg.ends_with("123)") {
+                errmsg = format!("{errmsg}. You may not be able to load evtx files when there are spaces in the directory path. Please enclose the path with double quotes and remove any trailing slash at the end of the path.");
+            }
             if stored_static.verbose_flag {
                 AlertMessage::alert(&errmsg).ok();
             }