+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 **Enhancements:**
 
-- Added error message to indicate that trailing backslashes should not be used in a directory path specification. (#1166) (@hitenkoku, thanks for the suggestion from @joswr1ght)
+- Added an error message to indicate that when you can't load evtx files in Windows due to specifying a directory path with spaces in it, you need to remove the trailing backslash. (#1166) (@hitenkoku, thanks for the suggestion from @joswr1ght)
 - Optimized the number of records to load at a time. (#1175) (@yamatosecurity)
 
 **Bug Fixes:**