+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 **Enhancements:**
 
+- Added an error message to indicate that when you can't load evtx files in Windows due to specifying a directory path with spaces in it, you need to remove the trailing backslash. (#1166) (@hitenkoku, thanks for the suggestion from @joswr1ght)
 - Optimized the number of records to load at a time. (#1175) (@yamatosecurity)
 
 **Bug Fixes:**